requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
idna==3.11
inflection==0.5.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
nltk==3.8.1
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdfium2==5.0.0
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
gunicorn==21.2.0
sentry-sdk==1.40.0